requirements.txt
streamlit
google-generativeai
PyMuPDF
streamlit-mic-recorder